@@ -1,7 +1,6 @@
 {
   "name": "alecture",
   "version": "1.0.0",
-  "description": "",
   "main": "index.js",
   "scripts": {
     "dev": "cross-env TS_NODE_PROJECT=\"tsconfig-for-webpack-config.json\" webpack serve --env development",
@@ -17,59 +16,61 @@
     "@loadable/component": "^5.14.1",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.4.3",
     "@types/gravatar": "^1.8.1",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.1",
+    "@types/react": "^17.0.44",
+    "@types/react-dom": "^17.0.15",
     "autosize": "^4.0.2",
     "axios": "^0.21.1",
     "cross-env": "^7.0.3",
-    "css-loader": "^5.0.2",
+    "css-loader": "^5.2.7",
     "dayjs": "^1.10.4",
     "gravatar": "^1.8.1",
-    "react": "^17.0.1",
+    "history": "^5.3.0",
+    "react": "^17.0.2",
     "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^17.0.1",
+    "react-dom": "^17.0.2",
     "react-mentions": "^4.1.1",
     "react-refresh": "^0.9.0",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-router": "^5.3.1",
+    "react-router-dom": "^6.3.0",
     "react-toastify": "^7.0.3",
     "regexify-string": "^1.0.5",
     "socket.io-client": "^2.4.0",
     "style-loader": "^2.0.0",
     "swr": "^0.4.2",
     "ts-node": "^9.1.1",
-    "typescript": "^4.1.5",
+    "typescript": "^4.6.3",
     "webpack-bundle-analyzer": "^4.4.0",
-    "webpack-cli": "^4.5.0"
+    "webpack-cli": "^4.9.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/preset-env": "^7.12.16",
-    "@babel/preset-react": "^7.12.13",
-    "@babel/preset-typescript": "^7.12.16",
+    "@babel/core": "^7.17.9",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/preset-typescript": "^7.16.7",
     "@types/autosize": "^3.0.7",
     "@types/loadable__component": "^5.13.3",
-    "@types/node": "^14.14.28",
+    "@types/node": "^14.18.13",
     "@types/react-custom-scrollbars": "^4.0.7",
     "@types/react-mentions": "^4.1.0",
-    "@types/react-router": "^5.1.11",
-    "@types/react-router-dom": "^5.1.7",
+    "@types/react-router": "^5.1.18",
+    "@types/react-router-dom": "^5.3.3",
     "@types/socket.io-client": "^1.4.35",
-    "@types/webpack": "^4.41.26",
+    "@types/webpack": "^4.41.32",
     "@types/webpack-bundle-analyzer": "^3.9.1",
-    "@types/webpack-dev-server": "^3.11.1",
-    "babel-loader": "^8.2.2",
-    "eslint": "^7.20.0",
+    "@types/webpack-dev-server": "^3.11.6",
+    "babel-loader": "^8.2.5",
+    "eslint": "^7.32.0",
     "eslint-config-prettier": "^7.2.0",
     "eslint-config-react-app": "^6.0.0",
     "eslint-plugin-flowtype": "^5.3.1",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-prettier": "^3.4.1",
     "eslint-plugin-react": "^7.22.0",
     "fork-ts-checker-webpack-plugin": "^6.1.0",
-    "prettier": "^2.2.1",
-    "webpack": "^5.22.0",
-    "webpack-dev-server": "^3.11.2"
-  }
+    "prettier": "^2.6.2",
+    "webpack": "^5.72.0",
+    "webpack-dev-server": "^3.11.3"
+  },
+  "description": ""
 }
