@@ -1,19 +1,18 @@
 import React from 'react';
-import loadable from '@loadable/component';
-import { Switch, Route, Redirect } from 'react-router-dom';
 
-const LogIn = loadable(() => import('@pages/LogIn'));
-const SignUp = loadable(() => import('@pages/SignUp'));
-const Workspace = loadable(() => import('@layouts/Workspace'));
+import { Routes, Route, Navigate } from 'react-router-dom';
+
+import LogIn from '@pages/LogIn';
+import SignUp from '@pages/SignUp';
+//import Workspace from '@layouts/Workspace';
 
 const App = () => {
   return (
-    <Switch>
-      <Redirect exact path="/" to="/login" />
-      <Route path="/login" component={LogIn} />
-      <Route path="/signup" component={SignUp} />
-      <Route path="/workspace/:workspace" component={Workspace} />
-    </Switch>
+    <Routes>
+      <Route path="/" element={<Navigate replace to="/login" />} />
+      <Route path="/login" element={<LogIn />} />
+      <Route path="/signup" element={<SignUp />} />
+    </Routes>
   );
 };
 
